+++ src/interfaces/IWETH.sol
@@ -4,5 +4,5 @@ import { IERC20 } from "@openzeppelin/token/ERC20/IERC20.sol";
 
 interface IWETH is IERC20 {
     function deposit() external payable;
-    function withdraw(uint) external;
+    function withdraw(uint256) external;
 }
